requirements.txt
numpy==1.19.5
opencv-python-headless==4.5.4.58
streamlit==1.1.0
Pillow==8.4.0
tensorflow-cpu==2.6.0